📝 rename runtime additional context types
