+++ packages/types/src/agent/index.ts
@@ -7,5 +7,5 @@ export * from './graph';
 export * from './item';
 export * from './modelSelection';
 export * from './pluginConfig';
-export * from './runtimeContext';
+export * from './runtimeAdditionalContext';
 export * from './tts';
